Reverted r20380 because it breaks all statements that print an OID sub-ID on any 64-bit platform. The effect of r20380 on big endian architectures was that a "0" would be printed instead of the OID sub-ID (at least with the LP64 and the ILP64 models).
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@20400 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/include/net-snmp/library/oid.h
@@ -6,7 +6,7 @@
 #endif
 
 #ifndef EIGHTBIT_SUBIDS
-typedef u_long oid;
+typedef uint32_t oid;
 #define MAX_SUBID   0xFFFFFFFF
 #define NETSNMP_PRIo NETSNMP_PRI32
 #else
